requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.2.2
SQLAlchemy==1.4.41
python-dotenv==0.21.0
authlib==1.3.2
requests==2.32.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
Faker==19.13.0
Flask-Migrate==4.0.5
python-dateutil==2.8.2
email-validator==2.1.0.post1
